Network.AWS.Budgets.CreateNotification

Description

Create a new Notification with subscribers for a budget

Creating a Request

createNotification

Arguments

:: Text
cnAccountId

-> Text
cnBudgetName

-> Notification
cnNotification

-> NonEmpty Subscriber
cnSubscribers

-> CreateNotification

Creates a value of CreateNotification with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cnAccountId - Undocumented member.

cnBudgetName - Undocumented member.

cnNotification - Undocumented member.

cnSubscribers - Undocumented member.

data CreateNotification

Request of CreateNotification

See: createNotification smart constructor.

Request Lenses

cnAccountId :: Lens' CreateNotification Text

Undocumented member.

cnBudgetName :: Lens' CreateNotification Text

Undocumented member.

cnNotification :: Lens' CreateNotification Notification

Undocumented member.

cnSubscribers :: Lens' CreateNotification (NonEmpty Subscriber)

Undocumented member.

Destructuring the Response

createNotificationResponse

Arguments

:: Int
cnrsResponseStatus

-> CreateNotificationResponse

Creates a value of CreateNotificationResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cnrsResponseStatus - -- | The response status code.

data CreateNotificationResponse

Response of CreateNotification

See: createNotificationResponse smart constructor.

Response Lenses

cnrsResponseStatus :: Lens' CreateNotificationResponse Int

- | The response status code.