Network.AWS.Budgets.DescribeBudgets

Description

Get all budgets for an account

Creating a Request

describeBudgets

Arguments

:: Text
dbAccountId

-> DescribeBudgets

Creates a value of DescribeBudgets with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dbNextToken - Undocumented member.

dbMaxResults - Undocumented member.

dbAccountId - Undocumented member.

data DescribeBudgets

Request of DescribeBudgets

See: describeBudgets smart constructor.

Request Lenses

dbNextToken :: Lens' DescribeBudgets (Maybe Text)

Undocumented member.

dbMaxResults :: Lens' DescribeBudgets (Maybe Natural)

Undocumented member.

dbAccountId :: Lens' DescribeBudgets Text

Undocumented member.

Destructuring the Response

describeBudgetsResponse

Arguments

:: Int
dbrsResponseStatus

-> DescribeBudgetsResponse

Creates a value of DescribeBudgetsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dbrsNextToken - Undocumented member.

dbrsBudgets - Undocumented member.

dbrsResponseStatus - -- | The response status code.

data DescribeBudgetsResponse

Response of DescribeBudgets

See: describeBudgetsResponse smart constructor.

Response Lenses

dbrsNextToken :: Lens' DescribeBudgetsResponse (Maybe Text)

Undocumented member.

dbrsBudgets :: Lens' DescribeBudgetsResponse [Budget]

Undocumented member.

dbrsResponseStatus :: Lens' DescribeBudgetsResponse Int

- | The response status code.