Network.AWS.Budgets.DescribeSubscribersForNotification

Description

Get subscribers of a notification

Creating a Request

describeSubscribersForNotification

Arguments

:: Text
dsfnAccountId

-> Text
dsfnBudgetName

-> Notification
dsfnNotification

-> DescribeSubscribersForNotification

Creates a value of DescribeSubscribersForNotification with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dsfnNextToken - Undocumented member.

dsfnMaxResults - Undocumented member.

dsfnAccountId - Undocumented member.

dsfnBudgetName - Undocumented member.

dsfnNotification - Undocumented member.

data DescribeSubscribersForNotification

Request of DescribeSubscribersForNotification

See: describeSubscribersForNotification smart constructor.

Request Lenses

dsfnNextToken :: Lens' DescribeSubscribersForNotification (Maybe Text)

Undocumented member.

dsfnMaxResults :: Lens' DescribeSubscribersForNotification (Maybe Natural)

Undocumented member.

dsfnAccountId :: Lens' DescribeSubscribersForNotification Text

Undocumented member.

dsfnBudgetName :: Lens' DescribeSubscribersForNotification Text

Undocumented member.

dsfnNotification :: Lens' DescribeSubscribersForNotification Notification

Undocumented member.

Destructuring the Response

describeSubscribersForNotificationResponse

Arguments

:: Int
dsfnrsResponseStatus

-> DescribeSubscribersForNotificationResponse

Creates a value of DescribeSubscribersForNotificationResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dsfnrsNextToken - Undocumented member.

dsfnrsSubscribers - Undocumented member.

dsfnrsResponseStatus - -- | The response status code.

data DescribeSubscribersForNotificationResponse

Response of DescribeSubscribersForNotification

See: describeSubscribersForNotificationResponse smart constructor.

Response Lenses

dsfnrsNextToken :: Lens' DescribeSubscribersForNotificationResponse (Maybe Text)

Undocumented member.

dsfnrsSubscribers :: Lens' DescribeSubscribersForNotificationResponse (Maybe (NonEmpty Subscriber))

Undocumented member.

dsfnrsResponseStatus :: Lens' DescribeSubscribersForNotificationResponse Int

- | The response status code.