Network.AWS.Budgets.Types

Description

Service Configuration

budgets :: Service

API version 2016-10-20 of the Amazon Budgets SDK configuration.

Errors

_InvalidParameterException :: AsError a => Getting (First ServiceError) a ServiceError

This exception is thrown if any request is given an invalid parameter. E.g., if a required Date field is null.

_InternalErrorException :: AsError a => Getting (First ServiceError) a ServiceError

This exception is thrown on an unknown internal failure.

_ExpiredNextTokenException :: AsError a => Getting (First ServiceError) a ServiceError

This exception is thrown if the paging token is expired - past its TTL

_NotFoundException :: AsError a => Getting (First ServiceError) a ServiceError

This exception is thrown if a requested entity is not found. E.g., if a budget id doesn't exist for an account ID.

_InvalidNextTokenException :: AsError a => Getting (First ServiceError) a ServiceError

This exception is thrown if paging token signature didn't match the token, or the paging token isn't for this request

_DuplicateRecordException :: AsError a => Getting (First ServiceError) a ServiceError

The exception is thrown when customer tries to create a record (e.g. budget) that already exists.

_CreationLimitExceededException :: AsError a => Getting (First ServiceError) a ServiceError

The exception is thrown when customer tries to create a record (e.g. budget), but the number this record already exceeds the limitation.

BudgetType

data BudgetType

The type of a budget. It should be COST, USAGE, or RI_UTILIZATION.

Constructors

Cost

RiUtilization

Usage

ComparisonOperator

data ComparisonOperator

The comparison operator of a notification. Currently we support less than, equal to and greater than.

Constructors

EqualTo

GreaterThan

LessThan

NotificationType

data NotificationType

The type of a notification. It should be ACTUAL or FORECASTED.

Constructors

Actual

Forecasted

SubscriptionType

data SubscriptionType

The subscription type of the subscriber. It can be SMS or EMAIL.

Constructors

Email

SNS

TimeUnit

data TimeUnit

The time unit of the budget. e.g. MONTHLY, QUARTERLY, etc.

Constructors

Annually

Daily

Monthly

Quarterly

Budget

data Budget

AWS Budget model

See: budget smart constructor.

budget

Arguments

:: Text
bBudgetName

-> Spend
bBudgetLimit

-> CostTypes
bCostTypes

-> TimeUnit
bTimeUnit

-> TimePeriod
bTimePeriod

-> BudgetType
bBudgetType

-> Budget

Creates a value of Budget with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

bCalculatedSpend - Undocumented member.

bCostFilters - Undocumented member.

bBudgetName - Undocumented member.

bBudgetLimit - Undocumented member.

bCostTypes - Undocumented member.

bTimeUnit - Undocumented member.

bTimePeriod - Undocumented member.

bBudgetType - Undocumented member.

bCalculatedSpend :: Lens' Budget (Maybe CalculatedSpend)

Undocumented member.

bCostFilters :: Lens' Budget (HashMap Text [Text])

Undocumented member.

bBudgetName :: Lens' Budget Text

Undocumented member.

bBudgetLimit :: Lens' Budget Spend

Undocumented member.

bCostTypes :: Lens' Budget CostTypes

Undocumented member.

bTimeUnit :: Lens' Budget TimeUnit

Undocumented member.

bTimePeriod :: Lens' Budget TimePeriod

Undocumented member.

bBudgetType :: Lens' Budget BudgetType

Undocumented member.

CalculatedSpend

data CalculatedSpend

A structure holds the actual and forecasted spend for a budget.

See: calculatedSpend smart constructor.

calculatedSpend

Arguments

:: Spend
csActualSpend

-> CalculatedSpend

Creates a value of CalculatedSpend with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

csForecastedSpend - Undocumented member.

csActualSpend - Undocumented member.

csForecastedSpend :: Lens' CalculatedSpend (Maybe Spend)

Undocumented member.

csActualSpend :: Lens' CalculatedSpend Spend

Undocumented member.

CostTypes

data CostTypes

This includes the options for getting the cost of a budget.

See: costTypes smart constructor.

costTypes

Arguments

:: Bool
ctIncludeTax

-> Bool
ctIncludeSubscription

-> Bool
ctUseBlended

-> CostTypes

Creates a value of CostTypes with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ctIncludeTax - Undocumented member.

ctIncludeSubscription - Undocumented member.

ctUseBlended - Undocumented member.

ctIncludeTax :: Lens' CostTypes Bool

Undocumented member.

ctIncludeSubscription :: Lens' CostTypes Bool

Undocumented member.

ctUseBlended :: Lens' CostTypes Bool

Undocumented member.

Notification

data Notification

Notification model. Each budget may contain multiple notifications with different settings.

See: notification smart constructor.

notification

Arguments

:: NotificationType
nNotificationType

-> ComparisonOperator
nComparisonOperator

-> Double
nThreshold

-> Notification

Creates a value of Notification with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

nNotificationType - Undocumented member.

nComparisonOperator - Undocumented member.

nThreshold - Undocumented member.

nNotificationType :: Lens' Notification NotificationType

Undocumented member.

nComparisonOperator :: Lens' Notification ComparisonOperator

Undocumented member.

nThreshold :: Lens' Notification Double

Undocumented member.

NotificationWithSubscribers

data NotificationWithSubscribers

A structure to relate notification and a list of subscribers who belong to the notification.

See: notificationWithSubscribers smart constructor.

notificationWithSubscribers

Arguments

:: Notification
nwsNotification

-> NonEmpty Subscriber
nwsSubscribers

-> NotificationWithSubscribers

Creates a value of NotificationWithSubscribers with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

nwsNotification - Undocumented member.

nwsSubscribers - Undocumented member.

nwsNotification :: Lens' NotificationWithSubscribers Notification

Undocumented member.

nwsSubscribers :: Lens' NotificationWithSubscribers (NonEmpty Subscriber)

Undocumented member.

Spend

data Spend

A structure represent either a cost spend or usage spend. Contains an amount and a unit.

See: spend smart constructor.

spend

Arguments

:: Text
sAmount

-> Text
sUnit

-> Spend

Creates a value of Spend with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

sAmount - Undocumented member.

sUnit - Undocumented member.

sAmount :: Lens' Spend Text

Undocumented member.

sUnit :: Lens' Spend Text

Undocumented member.

Subscriber

data Subscriber

Subscriber model. Each notification may contain multiple subscribers with different addresses.

See: subscriber smart constructor.

subscriber

Arguments

:: SubscriptionType
sSubscriptionType

-> Text
sAddress

-> Subscriber

Creates a value of Subscriber with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

sSubscriptionType - Undocumented member.

sAddress - Undocumented member.

sSubscriptionType :: Lens' Subscriber SubscriptionType

Undocumented member.

sAddress :: Lens' Subscriber Text

Undocumented member.

TimePeriod

data TimePeriod

A time period indicated the start date and end date of a budget.

See: timePeriod smart constructor.

timePeriod

Arguments

:: UTCTime
tpStart

-> UTCTime
tpEnd

-> TimePeriod

Creates a value of TimePeriod with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

tpStart - Undocumented member.

tpEnd - Undocumented member.

tpStart :: Lens' TimePeriod UTCTime

Undocumented member.

tpEnd :: Lens' TimePeriod UTCTime

Undocumented member.