Network.AWS.Budgets.UpdateNotification

Description

Update the information about a notification already created

Creating a Request

updateNotification

Arguments

:: Text
unAccountId

-> Text
unBudgetName

-> Notification
unOldNotification

-> Notification
unNewNotification

-> UpdateNotification

Creates a value of UpdateNotification with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

unAccountId - Undocumented member.

unBudgetName - Undocumented member.

unOldNotification - Undocumented member.

unNewNotification - Undocumented member.

data UpdateNotification

Request of UpdateNotification

See: updateNotification smart constructor.

Request Lenses

unAccountId :: Lens' UpdateNotification Text

Undocumented member.

unBudgetName :: Lens' UpdateNotification Text

Undocumented member.

unOldNotification :: Lens' UpdateNotification Notification

Undocumented member.

unNewNotification :: Lens' UpdateNotification Notification

Undocumented member.

Destructuring the Response

updateNotificationResponse

Arguments

:: Int
unrsResponseStatus

-> UpdateNotificationResponse

Creates a value of UpdateNotificationResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

unrsResponseStatus - -- | The response status code.

data UpdateNotificationResponse

Response of UpdateNotification

See: updateNotificationResponse smart constructor.

Response Lenses

unrsResponseStatus :: Lens' UpdateNotificationResponse Int

- | The response status code.